Support multiple functions with the same name defined in the same file
The callgrind format does not have a native way of representing multiple
functions with the same name defined in the same file.  Such function are common
in object-oriented code, where multiple classes can have methods with the same
name.  We de-duplicate such methods by appending the line number they are
defined on to the function name when there is a name collision.

Prior to this commit, functions with the same name defined in the same file were
aggregated together in the resulting callgrind profile.