Change units from ticks (milliseconds) to nanoseconds
For profiles where many functions take less than 1 millisecond, the total time
calculated via _summary() (the max total time) and the total time that
kcachegrind calculates (the sum of the exclusive times) can be very large.  This
leads to kcachegrind showing, e.g., that the most expensive function takes 220%
of the total time.  Switching to nanosecond precision largely mitigates this and
has the advantage of making less expensive functions show up in the profile
(functions with a cost of 0 appear to be dropped).